backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-axes==8.3.1
django-cleanup==9.0.0
django-cors-headers==4.9.0
djangorestframework==3.17.1
pillow==12.2.0
python-decouple==3.8
sqlparse==0.5.5
